Update product details ui part.
