+++ ecommers/lib/core/provider_models/index.dart
@@ -5,7 +5,6 @@ export 'forgot_password_provider_model.dart';
 export 'home_provider_model.dart';
 export 'log_in_provider_model.dart';
 export 'product_page_provider_model.dart';
-export 'product_page_tabs_provider_model.dart';
 export 'product_tab_provider_model.dart';
 export 'products_grid_provder_model.dart';
 export 'provider_model_base.dart';
