+++ ecommers/lib/ui/widgets/product_page/index.dart
@@ -1,6 +1,5 @@
 export 'details_tab.dart';
 export 'header.dart';
 export 'product_page_bottom_view.dart';
-export 'product_page_tabs_view.dart';
 export 'product_tab.dart';
 export 'reviews_tab.dart';
