+++ ecommers/lib/ui/widgets/product_page/product_tab.dart
@@ -35,15 +35,14 @@ class ProductTab extends StatelessWidget {
       ),
       child: Consumer<ProductTabProviderModel>(
         builder: (context, ProductTabProviderModel model, child) {
-          return Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
+          return ListView(
             children: [
               Text(
                 localization.selectColor,
                 style: Theme.of(context).textTheme.subtitle2,
               ),
               SizedBox(
-                height: 100,
+                height: Insets.x25,
                 child: _createColorsList(noInfoTextStyle, model, localization),
               ),
               Text(
@@ -52,7 +51,7 @@ class ProductTab extends StatelessWidget {
               ),
               const SizedBox(height: Insets.x5),
               SizedBox(
-                height: 50,
+                height: Insets.x12_5,
                 child: _createSizesList(noInfoTextStyle, model, localization),
               ),
             ],
@@ -125,7 +124,7 @@ class ProductTab extends StatelessWidget {
           return GestureDetector(
             onTap: () => _selectColor(i, model),
             child: Container(
-              width: 50,
+              width: Insets.x12_5,
               decoration: BoxDecoration(
                 color: Color(model.colors[i].color),
                 shape: BoxShape.circle,
